I refactored everything I'd done on the graph building and DFS portions of the project under the graph.java file. In this version, the Graph class is used to represent a graph with an adjacency matrix, an array of 'visited' values, along with some other attributes like title, found status, pathcost and search path.
All of the methods are now private methods to be used on objects of the 
Graph class, rather than static methods. I think this makes for cleaner, 
more readable code as well as simplifying the process of tracking each 
of the attributes in the context of a particular graph.

I did this in response to Granit's suggestion for a global boolean 
value. I think it worked out fairly well.

I also refactored DFS so that before it iterates through all of the 
nodes that may be connected to the 'current' node, it checks to see if 
the target node (or accept state) has a connecting edge. 

At the moment. The DFS almost traverses the graph in the expected 
manner: For example, if searching for any node from node 3(which has 
edges connecting it to all of the other nodes) the unweighted path cost 
is always 1 and the search path is 3->1, 3->2, etc.

The current snag is that if the target node is not directly connected to 
the start node (like searching for node 6 from node 1) it goes through 
the expected path, but then goes back to node 1 a couple of times. I 
realize this is an issue pertaining to one of the inner branches of the 
method, but I haven't quite figured it out yet. A pair of fresh eyes 
would probably go a long way.