OK, DFS works as it should now.
Given the problem of traveling from one specified node to another 
specified node, without ever visiting the same node twice, this DFS 
algorithm often finds its mark. It searches the graph depth-first, with 
a "left-side" or "less-than" branch preference. Since it does not 
revisit a node it has already visited on that search attempt, there are 
graphs for which such a path exists that will not be found by this 
algorithm.

It has been tested agaisnt 6 graphs, using every valid combination of 
start and accept states for each graph. For those graphs, I believe that 
it has successfully found a path any time it is possible.